+++ src/main/java/com/game/core/comment/application/CommentService.java
@@ -1,16 +1,16 @@
 package com.game.core.comment.application;
 
 import com.game.core.board.domain.Board;
-import com.game.core.board.domain.vo.LikeTag;
 import com.game.core.board.infrastructure.BoardRepository;
 import com.game.core.comment.dto.request.CreateCommentRequest;
 import com.game.core.comment.dto.request.UpdateCommentRequest;
-import com.game.core.comment.dto.response.ReadCommentRequest;
+import com.game.core.comment.dto.response.ReadCommentResponse;
 import com.game.core.comment.infrastructure.CommentRepository;
 import com.game.core.error.dto.ErrorMessage;
 import com.game.core.comment.domain.Comment;
 import com.game.core.error.exception.board.NotFindBoardException;
 import com.game.core.member.dto.LoggedInMember;
+import com.game.core.member.infrastructure.UserRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -26,7 +26,7 @@ public class CommentService {
     private final BoardRepository boardRepository;
 
     @Transactional
-    public void createComment(Long id, CreateCommentRequest createCommentRequest){
+    public void createComment(Long id,LoggedInMember loggedInMember, CreateCommentRequest createCommentRequest){
         Board board = boardRepository.findById(id)
             .orElseThrow(()-> {
                 throw new NotFindBoardException(ErrorMessage.NOT_FIND_ID_BOARD);
@@ -34,6 +34,7 @@ public class CommentService {
         board.addComment(commentRepository.save(
             Comment.builder()
                 .content(createCommentRequest.getContent())
+                .userId(loggedInMember.getId())
                 .likeViews(board.getLikeCount())
                 .board(board)
                 .build()
@@ -87,27 +88,42 @@ public class CommentService {
 
 
     @Transactional
-    public List<ReadCommentRequest> getComments(Long boardId) {
+    public List<ReadCommentResponse> getComments(Long boardId, LoggedInMember loggedInMember) {
         List<Comment> comments = commentRepository.findAllByBoardId(boardId);
-        List<ReadCommentRequest> commentRequests = new ArrayList<>();
+        List<ReadCommentResponse> commentRequests = new ArrayList<>();
+        List<ReadCommentResponse> commentResults = new ArrayList<>();
+
 
         comments.forEach(s -> commentRequests.add(
-            ReadCommentRequest.builder()
+            ReadCommentResponse.builder()
                 .id(s.getId())
                 .boardId(s.getBoard().getId())
+                .userId(loggedInMember.getId())
                 .comment(s.getContent())
+                .mine(false)
                 .likeView(s.getLikeViews())
                 .build()
         ));
-    return commentRequests;
+        for(ReadCommentResponse x : commentRequests){
+            if(loggedInMember.getId().equals(x.getUserId())){
+                x.setMine(true);
+                commentResults.add(x);
+            }else{
+                commentResults.add(x);
+            }
+        }
+
+        return commentResults;
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, LoggedInMember loggedInMember) {
         Comment comment = commentRepository.findById(commentId)
             .orElseThrow(()-> {
-            throw new NotFindBoardException(ErrorMessage.NOT_FIND_ID_BOARD);
-        });
-        commentRepository.deleteById(comment.getId());
+                throw new NotFindBoardException(ErrorMessage.NOT_FIND_ID_BOARD);
+            });
+        if(Objects.equals(loggedInMember.getId(), comment.getBoard().getUserName())){
+            commentRepository.deleteById(comment.getId());
+        }
     }
 }
